Remove Implicit_Dereference from Smart_Ptr
The only way to use Implicit_Dereference is if the Smart_Ptr type has
a discriminant. As the Smart_Ptr type has to be a Controlled, and
therefore tagged, type, this means the discriminant cannot be changed
which means a Smart_Ptr cannot be reassigned. This makes them less than
useful. Instead there are new P and Get operations to return a reference
type and the underlying pointer respectively.