Remove duplicate pointer E from Unique_Ptr
Unique_Ptr has to take a general access type discriminant in order to
allow use of Implicit_Dereference. However we need to be more
restrictive about the target objects to ensure they can be validly
deallocated and are not stack or static variables. Previously two copies
of the pointer were kept but as we know that the type can only be set up
via a named pointer type it is safe to do unchecked access type
conversions.