Actually make Weak_Ptr finalization idempotent
The previous check for the Counter_Ptr being null before decrementing
counters and (possibly) destroying the Counter was not sufficient. If
there was an existing Smart_Ptr pointing to the object the related
Counter will never be destroyed, but the Weak_Ptr finalization routine
could double-decrement the counter in the rare occasion when a
finalization routine is called twice on one object. Now we force the
pointer to the Counter to null regardless of whether it is destroyed,
so we will not double-decrement.