Smart_Ref does not need an Invalid flag
Reading the AARM 7.6 Ramification 2.b/2 the Initialization procedure for
Controlled types is only called if the value was not created from an
aggregate - as they are by the Make_xxx_xxx functions. For the Smart_Ref
Initialization procedure an immediate 'raise' prevents any default
initialization as desired. There is no need for the Invalid boolean, and
removing it saves 8 bytes per Smart_Ref (including padding on a 64-bit
machine).