Move Element from Counters_Spec to Weak_Ptr
The counters should not have a pointer to what they are counting. This
allows the counters to be more re-usable outside of the context of
Smart_Ptrs. The only use of Element was to recreate Smart_Ptr or
Smart_Ref from the Counter held by Weak_Ptr, but it would seem at least
as appropriate for Weak_Ptr to possess this information for itself,
rather than relying on the Counter.