Simplify unique_ptrs
As with commit #0202e83f, the Initialize routines for Unique_Ptr and
Unique_Const_Ptr can simply raise an exception without needing to check
the Invalid flag. However in this case we cannot get rid of the Invalid
flag. Per the standards Finalize has to be idempotent so we need to
prevent double deallocation if it is called twice on the same object. In
the Smart_Ptrs types we can null out the pointers to indicate that no
further deallocation is appropriate, but because we are using access
discriminants here, this does not work.