Avoid testing elements when finalizing a ref
When a Flyweight.Refcounted_Element_Ref is being finalized, it is
necessary to find the associated Element_Access in the relevant bucket
of the Flyweight. The candidates can be tested by comparing pointers.
There should be no duplicate element values in a Flyweight so there is
no difference between comparing element identity and value.