Split out the hash-table from the reference type
There is now a separation into three layers. The top layer defines the
reference type and depends on the second layer, which defines the
hash-table type. This in turn depends on the third layer, which defines
the basic linked-list type held in the hash-table buckets. The third
layer also takes responsibility for allocating and releasing resources.

This results in a further dependency, in that a reference-counting
reference type in the top layer can only be properly paired with a
reference-counting resource management implementation in the third
layer.