Fix bug when inserting an existing resource
The Insert routine was only checking value equality, and if found was
automatically deallocating the value passed in via access parameter.
This is bad if the access parameter passed was actually already inside
the Flyweight itself!. Now we check for exact access value equality
first and only check value equality (and potentially deallocate the
parameter passed in) if that fails.