Add protected hashtables
A simple job given the previous work on modelling the dependencies
between parts of the system. Currently the protected Flyweight is a
single protected object. Potentially having a separate protected object
barrier for each individual hash-table bucket would be faster if there
was contention between tasks to add/remove resources, but this is
not likely to be a problem in practice and does not justify the
additional complexity / resource usage.