Change some tests to Asserts
Some run-time checks for null access values etc have been changed to
pragma Asserts. This has only been done where a failure of the assertion
implies that the relevant type's internal state has become corrupted,
which implies an error somewhere in the library code. Depending on the
compiler's assertion policy these can be removed in production builds.

The tests have not been changed where failure implies an error in the
code of the user of the library. For example, attempting to make a
Smart_Ref (which can't be null) from a null Smart_Ptr (which is a valid
Smart_Ptr value) is a user error, not evidence of a fault in the
Smart_Ptrs package code. This is still an explicit test.