Fix unit tests for optimized builds
Some of the unit tests are checking for the presence of appropriate
preconditions and assertions. However in optimized builds these will be
removed by the compiler, so the test code will either trigger false
failure notifications or will segfault. These tests are now only run if
the build has assertions enabled.