Fix scope issues in unique_ptrs
In order to have the syntax sugar of implicit dereferencing using
the Implicit_Deference attribute, the Unique_Pointer types have to have
an access discriminant. This triggers run-time access scope checks when
these checks are enabled in the compiler, even when Unchecked_Conversion
is used (at least in recent versions of GNAT). The only solution I have
found so far is to store the pointer twice, once as an anonymous access
discriminant and once as the pointer type defined in the package.

Signed-off-by: James Humphry <git@binecho.co.uk>