Fix Smart_Ref type for new GNAT rules on Unchecked_Conversion
As with Unique_Ptrs, the later versions of GNAT do not allow
Unchecked_Conversion to bypass access value scope checks. This means
that it is not possible to use the anonymous access type as the
discriminant of the Smart_Ref record (which allows implicit dereference)
to also be used for things such as deallocating the storage. The only
solution so far is to store the access value again inside the record,
using a named access type.

Signed-off-by: James Humphry <git@binecho.co.uk>