Fix flyweights and kvflyweights for new rules on Unchecked_Conversion
As with the previous commits, it is not longer possible to create
reference types which store an access value as an access discriminant
(for implicit dereferencing) but also convert that anonymous access
discriminant for use when needed. Unchecked_Conversion on access types
no longer ignores the access scope. It is therefore necessary to store
the access value twice in the reference.

Signed-off-by: James Humphry <git@binecho.co.uk>